% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_learning_curve.R
\name{estimate_learning_curve}
\alias{estimate_learning_curve}
\alias{estimate_learning_cuve}
\title{Estimate learning curve}
\usage{
estimate_learning_curve(output, sample_size, chance = 0.05)
}
\arguments{
\item{output}{Takes the dataframe from \code{train_pred_model_on_subsets()} to perform an estimation of the learning curve.}

\item{sample_size}{An integer defining up to which sample size the errorrate should be predicted.}

\item{chance}{A double defining what the maximum pvalue for the errorate that are used should be (Default: chance = 0.05)}
}
\value{
A dataframe containing the samplesizes which were predicted and the predicted mean and 25/75 \\% quantile errorates for these samplesizes.
}
\description{
This function takes the ouput form \code{train_pred_model_on_subsets()} and a sample size up to which the errorates should be predicted, to be later displayed with \code{plot_learning_curve()}.
}
\seealso{
\code{train_pred_model_on_subsets()}, \code{plot_learning_curve()}.
}
\author{
Jonathan Krause
}
